Signout API

+++ client_admin/src/actions/auth.actions.js
@@ -49,11 +49,22 @@ export const isUserLoggedIn = () => {
     }
 }
 
+
+
 export const signout = () => {
     return async dispatch => {
-        localStorage.clear();
-        dispatch({
-            type: authConstants.LOGOUT_REQUEST
-        })
+        
+        dispatch({ type: authConstants.LOGOUT_REQUEST })
+        const res = await axios.post(`/admin/signout`)
+
+        if (res.status === 200) {
+            localStorage.clear();
+        dispatch({ type: authConstants.LOGOUT_SUCCESS })
+        } else {
+            dispatch({ 
+                type: authConstants.LOGOUT_SUCCESS, 
+                payload: { error: res.data.error }
+            })
+        }
     }
 }
